added assertion block modules
